refactor(ekf2): move SensorSlotBinder to a shared library
Allow modules other than ekf2 to reuse the device-ID based parameter
slot binding, in particular the sensors module for the per-sensor
SENS_FLOW{i}_* parameters.
