+++ src/lib/sensor_slot_binder/SensorSlotBinder.hpp
